Mod Update
Lightsabers are functioning (they will need a model remake for the blad)
Blasters are functioning.
For Next Update:
CustomPlayerRenderer for armor models, and for custom player animations.
Fix sound's not loading.
Begin adding the rest of the weapons and armor so that a test server can
be up soon.
After that, work on Gamemodes!